backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
pdfplumber
python-docx
requests
aiofiles
